+++ src/main/java/com/kube/noon/common/messagesender/NotificationMessageSender.java
@@ -1,9 +1,9 @@
-package com.kube.noon.notification.service.sender;
+package com.kube.noon.common.messagesender;
 
 import com.kube.noon.member.domain.Member;
 import com.kube.noon.notification.domain.NotificationType;
 
-public interface NotificationTransmissionAgent {
+public interface NotificationMessageSender extends MessageSender {
 
     public void send(Member receiver, String text, NotificationType notificationType);
 }
